和vue-cli同步

+++ server/models/user.js
@@ -1,6 +1,7 @@
 const sequelize = require('./db')
 const Sequelize = require('sequelize')
 const md5 = require('md5');
+const config = require("../configs/index.js")
 
 const User = sequelize.define('user', {
     user_name: {
@@ -11,20 +12,23 @@ const User = sequelize.define('user', {
     },
     user_phone:{
         type: Sequelize.STRING
+    },
+    user_type:{
+        type: Sequelize.INTEGER
     }
 });
-User.sync({ force: true }).then(function () {
-    // 已创建数据表
-    console.log("已经创建user表")
-    // let pwd = md5("superAdmin")
-    // User.create({
-    //     user_name: "superAdmin",
-    //     user_pwd: pwd,
-    //     user_phone: "1234567890",
-    //     user_type: 9999,
-    //     user_time: "0"
-    // })
+if(config.mysql.reInit){
+    User.sync({ force: true }).then(function () {
+        let pwd = md5(config.admin.password)
+        User.create({
+            user_name: config.admin.username,
+            user_pwd: pwd,
+            user_phone: "1234567890",
+            user_type: 99,
+        })
+    
+    });
+}
 
-});
 
 module.exports = User;
